Learning-Journal-Week-7
This weeks learning journal includes improvements to my code so that it works on multiple annotations. First successfully using the hypothes.is api to save information for multiple annotations. Then using awk commands, an until loop, and an incrementally increasing variable to get information from specific sections of the annotation information. Next using grep to count the number of different annotations within the information and using that number to change the variable which controls when the until loop stops. Then cleaning the information so that each annotation is separated by a blank line, separating each section with a | and removing some of the extra characters.