Learning-Journal-Week-12
Additions to this learning journal involve progress on unit testing and the static website. In terms of unit testing I was still struggling with it, so I went to consultation hours with Brian to figure it out. He helped me install bats correctly, write a test for the wget command and run the test successfully. In terms of the static website, I created a static website with github pages which contains my deployment, execution instructions, and example code. This should satisfy the HD requirements of both documentation and quality assurance.